drivers: i2c: stm32: Factorize devices instantiation
Factorize the macros and definitions of the I2C devices resources
between I2C STM32 non-RTIO and RTIO drivers that are the same
aside resource that can be conditioned to CONFIG_I2C_RTIO being enabled.
This change requires I2C API structure (renamed here i2c_stm32_driver_api)
and i2c_stm32_init() initialization function are no more locally defined.
